fixed restock mode on DSM/Palace (#310)
* fixed restock mode on DSM/Palace

* fixed dsm please pray to god

* Refactor Get Checkout method to frontend

This commit refactors the get checkout method from the base checkout class to the frontend checkout class where it is actually needed. This means that the api checkout method will not need to perform an extra request that is unnecessary, but the frontend checkout will perform the necessary extra request.