Better Request Error Handling (#330)
* initial pass through error handling

* change thread pool size

* Refactor Error Handling to Separate Util Function

This commit creates a stateForError function that takes a given
error and returns a nextState for common errors. This includes
status code errors as well as request errors. This method is now
used in the error handlers for most request code to provide better
error handling for common issues.

If no state can be determined the function returns null and callers
can then process specific errors further.