Implement DsmUkParser
This commit implements a specific DsmParser variant for the UK Site. This
includes parsing for a custom.js sourced script in both the initial page
and product pages head element. The script is then fetched and its contents
are parsed to find the included hash property. If the hash property is
found on the initial page, the parsing for the hash property on the product
page is skipped -- this prevents an unnecessary additional request.