Update Special Parser Selections
This commit updates the getSpecialParser method to include the
refactored DsmParser as well as the new variants for specific sites.

The specialParser is also updated slightly to allow for asynchronous
implementations when parsing the initial/product pages. This allows the
DsmUkParser to function properly since it has to fetch the custom.js
file for parsing the hash property.