Refactor Hash Property Parsing for DSM Variants (#334)
* Refactor DsmParser

This commit refactors DsmParser to offer extensibility when parsing for
site/product specific hash properties. By default no parsing is done, but
subclasses of DsmParser can now provide an implementation to have this
parsing done if necessary on the initial page, product pages, or both

Additionally, the file is moved to a new folder where all dsm parser
variants can live.

* Implement DsmUsParser

This commit implements a specific DsmParser variant for the US Site. This
includes parsing a product page for a specific hash property that is
within a binary encoded string in a script attached to the product page.

* Implement DsmUkParser

This commit implements a specific DsmParser variant for the UK Site. This
includes parsing for a custom.js sourced script in both the initial page
and product pages head element. The script is then fetched and its contents
are parsed to find the included hash property. If the hash property is
found on the initial page, the parsing for the hash property on the product
page is skipped -- this prevents an unnecessary additional request.

* Update Special Parser Selections

This commit updates the getSpecialParser method to include the
refactored DsmParser as well as the new variants for specific sites.

The specialParser is also updated slightly to allow for asynchronous
implementations when parsing the initial/product pages. This allows the
DsmUkParser to function properly since it has to fetch the custom.js
file for parsing the hash property.

* Fix Task Runner Lints

This commit fixes some small lints to remove all lint errors/warnings
in the task runner package.

* Fix Typo in Logs

Co-Authored-By: pr1sm <dhanwada.dev@gmail.com>

* Finish Writing Doc Comment

This commit finishes an incomplete doc comment.