Use Proxy on CaptchaWindow instead of Protocol
This commit updates the method of proxying the captcha window to use a proxy instead of using electron's protocol api. This reduces the area of effect of the captcha server to just the captcha windows instead of having an effect on all requests.

Finally this commit removes a script tag to a file that doesn't exist in the captcha html.