Refactor Task Manager Wrapper to Enable Web Worker Thread in Future (#342)
* Remove Task Manager Wrapper

This commit removes the task manager wrapper file to accommodate the incoming TaskWrapper class. The new task will be responsible for all the same items as the TaskManagerWrapper, so the old class is no longer needed.

* Add TaskWrapper and TaskLauncher Window

This commit adds a new TaskWrapper class that launches a new hidden BrowserWindow that runs a TaskManagerWrapper class. Instead one file handling all responsibilities, these two classes handle managing task related functions.

The old TaskManagerWrapper is split between these two classes to effectively offload work done on the main process to a secondary process. This should hopefully improve responsiveness, but it mainly enables the future use of webworker threads for task runners.

This commit represents most of the work complete, but there are still a couple of lingering implementations to finish:
- Captcha Windows error out with invalid site key
- Cleanup for running task processes when quitting the app

* Rename Files and Misc Cleanup

This commit updates the naming of the refactored classes to better describe their function.

This commit also performs some small cleanup to other files.

* Add Task Cleanup

This commit adds the cleanup functionality for tasks when the main window closes. This prevents hanging tasks from continuing to run after the app quits.

* fix typo in comments

Co-Authored-By: pr1sm <dhanwada.dev@gmail.com>

* Pass correct sitekey

This commit fixes a bug that prevented the captcha sitekey from being set properly. a missing argument in the handler shifted parameters over, so the sitekey got set improperly.

* Use Proxy on CaptchaWindow instead of Protocol

This commit updates the method of proxying the captcha window to use a proxy instead of using electron's protocol api. This reduces the area of effect of the captcha server to just the captcha windows instead of having an effect on all requests.

Finally this commit removes a script tag to a file that doesn't exist in the captcha html.

* Rename IPC Channel

This commit changes an IPC channel name to be more descriptive.