Start Refactoring CaptchaWindowManager
This commit starts the refactor for the new CaptchaWindowManager. This manager will be in charge of all captcha windows instead of having a manager instance per window.

At this time, the old captcha window manager is still in use. Once the new manage is written. the regular WindowManager will have to be adjusted to use the new CaptchaWindowManager