Adjust TaskEvent Registration Logic
This commit adjusts the logic surrounding registering for task events.
Now the launcher will register for events as soon as the adapter is
ready. This prevents any possibility of the task event chain not being
set up properly to forward events to the frontend.

This commit also hides console logs so they only show in dev mode and
renames the manager.html file to a more appropriate launcher.html