Add Linter Exemptions for Certain Classes
We have a requirement to use instance methods within the Transformer subclasses, but the implementations don't use `this`. Since this is a valid exception to the rule, an eslint comment is added to the top of both to convey this.

See https://eslint.org/docs/rules/class-methods-use-this#exceptions for more details