Fix Delay and Webhook Event Passing
This commit updates the event handlers to allow responding to a bulk event. In this case the ID passed will be 'ALL'. This allows a single event to be emitted when adjusting a delay or webhook in the TaskManager for all tasks.

This also keeps the contract of purely using event based communication between the manager and runner.