Transition to Use New Captcha Window Manager
This commit replaces the old implementation of WindowManager to use the new CaptchaWindowManager for captcha/youtube functions.

The old captcha window manager implementation is removed. There is still some code in the WindowManager that needs to be moved over to the CaptchaWindowManager. For now, these calls are just proxied so they work.