Refactor Captcha Window Manager (#355)
* Start Refactoring CaptchaWindowManager

This commit starts the refactor for the new CaptchaWindowManager. This manager will be in charge of all captcha windows instead of having a manager instance per window.

At this time, the old captcha window manager is still in use. Once the new manage is written. the regular WindowManager will have to be adjusted to use the new CaptchaWindowManager

* Add Injectable Options for Youtube and Captcha Windows

This commit updates the create window functions for Youtube and Captcha to allow custom options and webPreferences to be injected in. This allows further customizations when creating these windows.

* More Implementation Progress

This commit adds more progress towards implementing a new CaptchaWindowManager.
- Specific functions to spawn youtube and captcha windows are added
- The end session handler is implemented.
- Adjustments to the Youtube Launch handler

* Transition to Use New Captcha Window Manager

This commit replaces the old implementation of WindowManager to use the new CaptchaWindowManager for captcha/youtube functions.

The old captcha window manager implementation is removed. There is still some code in the WindowManager that needs to be moved over to the CaptchaWindowManager. For now, these calls are just proxied so they work.

* Cleanup WindowManager

This commit adds some cleanup for the window manager.

The change theme handler has been commented out until #350 is addressed.

* Use Harvest State when Spawning New Captcha Window

This commit updates the Captcha Window Manager to automatically start harvesting on a new window if the rest of the windows are in the harvesting state. This allows you to open/close all captcha windows, but have new window harvest if a captcha is still required.

Issue: #106

* Increase Speed of Captcha Stopping

Instead of reloading the page, the captcha window now only loads the captcha methods once and resets when it stops harvesting. The captcha form is hidden to display the idle animation again. This prevents a reload cycle from happening everytime the captcha finishes harvesting

Issue: #272

* Remove Proxy on captcha window load

This commit removes the captcha server proxy when the window loads. This allows the youtube window experience to function properly.

* Fix Captcha Windows not closing on deactivate

This commit adds a call to always close the captcha windows when the main window closes. This should include both when closing and deactivating.

* Fix Captcha Window Auto Spawn

This commit fixes a bug where a captcha window would not be spawned if it was needed.

fixes #97
fixes #106
fixes #272
fixes #317