Captcha Autoclick Improvements (#374)
* Fix Mouse Event and Add more Events

This commit address two possible issues with autoclicking. The first is that the mouse event created does not include point information for where the mouse moves. We now use the given dest and source points when simulating an event.

The second is the number of elements to generate events for. This commit add a couple of more elements for the mouse to move through.

* Revert Explicit Loading

This commit reverts the explicit loading options from the initial call to the recaptcha api script. All options are attached to the container, which the api script will automatically use to render the widget.