Add Migrator Logic
This commit adds the main logic behind migrating a state from a given version to the latest. This requires the addition of the semver package to sort tracked versions sequentially.

This logic will not have to change, but new migrators will have to be imported as we make new changes to the state.