Add 0.0.0 Migrator
This commit adds the initial migrator that represents the current state tree. A copy of the initial state tree is added so the original source can be updated as changes are made. References to the original source files are added in comments. This base state will be used as an initial starting point for new migrations.