Add Migration Shortcut on Initial State
This commit adds a shortcut to the migration logic when the given state is null. In this case, we simply want to return the latest migrations initial state, so we don't need to go through the logic of creating a full length chain and going through meaningless migrations, ending up with the latest initial state.