Fix Windows Jest Caching Errors [temporary]
This commit temporarily fixes a bug with jest due to bad atomic file writing on Windows. This fix is a temporary on using the resolutions to force a lower version of the offending package.

A fix is available for this package (2.4.2) and is consumed in a newer version of jest (24.x), but both packages are depended on transiently through react-scripts, which has yet to consume the fixed version of jest. Until that fix makes its way to react-scripts, we can use this to pin the working version.