Source Initial State from latest migration version
This commit updates the frontend to source the initial state from the latest migration version. Instead of having the initial state parts defined in the definition files, the parts are exported similarly from several new files that pull the initial state from the migrator file. The migrator file in turn has a function that looks for the latest migrator and gets the initial state.

All import references were updated to reflect the change.