Loosen Ban Criteria for Proxies (#386)
* Adjust shouldBan flag to only ban on 403 errors

This commit loosens the should ban status to only being on 403 errors. This should prevent a chain of proxy bans from too many requests on the same site.

* Adjust Log Message

This commit updates the log message to add the number of proxies available for searching. This should help diagnose errors where the proxy list is getting out of sync.

* Fix typo

the `size` is a property of a `Map` not a function, so anytime we run this code, an error is thrown. Change this to a property call fixes the issue.

Co-Authored-By: pr1sm <dhanwada.dev@gmail.com>