Add Redux State Migration (#383)
* Add Migrator Logic

This commit adds the main logic behind migrating a state from a given version to the latest. This requires the addition of the semver package to sort tracked versions sequentially.

This logic will not have to change, but new migrators will have to be imported as we make new changes to the state.

* Add 0.0.0 Migrator

This commit adds the initial migrator that represents the current state tree. A copy of the initial state tree is added so the original source can be updated as changes are made. References to the original source files are added in comments. This base state will be used as an initial starting point for new migrations.

* Add 0.1.0 Migrator

This commit adds a 0.1.0 migrator to handle adding the first "versioned" state tree. This migrator simply attaches the version to the previous 0.0.0 state.

* Refactor Top Level Migrator

This commit refactors the top level migrator to allow a full migrator map to be injected instead of just the versions. This prevents a bug where a version is found (from injected versions), but doesn't exist in the migrator map, causing an error.

* Add Migration Shortcut on Initial State

This commit adds a shortcut to the migration logic when the given state is null. In this case, we simply want to return the latest migrations initial state, so we don't need to go through the logic of creating a full length chain and going through meaningless migrations, ending up with the latest initial state.

* Add Migrate State Action

This commit adds a global action to migrate the state tree and adds the call to the migrator in the top level reducer. Further. the initial state tree is updated to the latest tracked version (0.1.0).

* Call Migrate State

This commit updates the entry point of the frontend react to call migrate state.

* Fix typo

Co-Authored-By: pr1sm <dhanwada.dev@gmail.com>

* Fix Windows Jest Caching Errors [temporary]

This commit temporarily fixes a bug with jest due to bad atomic file writing on Windows. This fix is a temporary on using the resolutions to force a lower version of the offending package.

A fix is available for this package (2.4.2) and is consumed in a newer version of jest (24.x), but both packages are depended on transiently through react-scripts, which has yet to consume the fixed version of jest. Until that fix makes its way to react-scripts, we can use this to pin the working version.

* Adjust Directory Structure

This commit adjusts the directory structure to match the style of importing we have for actions and reducers. No functionality changes, just the file names, locations, and import paths.

* Add Migrator Logic Unit Tests

This commit adds a new suite of unit tests to focus on the testing the logic for the top level migrator. Test suite over specific migrators will follow.

* Add Unit Test Suites for Specific Migrators

This commit adds separate test suites for each migrator added.

* Finish adding Unit Tests

This commit updates the frontend to increase the coverage of changed/added code.

[Bonus]: Tests for the set theme action were also added.

* Source Initial State from latest migration version

This commit updates the frontend to source the initial state from the latest migration version. Instead of having the initial state parts defined in the definition files, the parts are exported similarly from several new files that pull the initial state from the migrator file. The migrator file in turn has a function that looks for the latest migrator and gets the initial state.

All import references were updated to reflect the change.

fixes #69