Remove All Previous Logs
To initially cleanup logs, we need to delete the log directory contents the _first_ time we run this changed version. The presence of the `audit.json` file is what we use to determine this.

If the log directory previously existed, without the audit file, we know we were using an old version, so we recursively delete the directory contents.

We can then proceed with all future files following the DRF cleanup model.