Add Better Log Cleanup (#390)
* Add Daily File Rotate Transport

This commit adds the daily file rotate transport to handle rotating files on a schedule. This will be used instead of the regular file transport.

* Use daily rotate file transport

This commit updates the logger implementation to use the new transport instead of the old file transport.

* Prevent Unnecessary Logs

This commit prevents unnecessary log messages from being printed. The content is unnecessary and drastically increases the size of the log file for little to no benefit.

* Remove All Previous Logs

To initially cleanup logs, we need to delete the log directory contents the _first_ time we run this changed version. The presence of the `audit.json` file is what we use to determine this.

If the log directory previously existed, without the audit file, we know we were using an old version, so we recursively delete the directory contents.

We can then proceed with all future files following the DRF cleanup model.

fixes #118