Ensure all parsers include product url
This commit updates the parser run methods to ensure the parser includes the product url. This ensures that restocking can happen without the need to search for the product again.

That product url is then stored as the restockUrl for the task's product -- this will allow restocking to have an easy reference to the endpoint it needs to poll for more information.