Add Restock Monitor (#392)
* switch endpoints to .js

* Split GenerateVariants to a Utility File

This commit splits a part of the monitor class out to a utility function. This will allow it to be used other classes as well without needing to extend Monitor.

* Add Initial Filter for Availability

This commit updates the generateVariants utility function to check for available variants before doing anything else. This allows any generated variants to be known as available, while also allowing us to return early if no variants are available.

* Add Variant Related Error Codes

This commit adds variant specific error codes and updates the generateVariants function throw errors with the relevant codes.

A new private function is added to the monitor to call the generateVariants function and handle the error returned (if it exists).

* Ensure all parsers include product url

This commit updates the parser run methods to ensure the parser includes the product url. This ensures that restocking can happen without the need to search for the product again.

That product url is then stored as the restockUrl for the task's product -- this will allow restocking to have an easy reference to the endpoint it needs to poll for more information.

* Consolidate Error Codes Map

This commit consolidates the error codes maps into one. Separate error maps added an unnecessary extra level.

Imports and usages have been updated to reflect this change.

* Implement Restock Monitor

This commit implements a sub class of the monitor that performs restock monitoring. If restocking is supported, the restock monitor will look for matched, available variants. If no variants are available, further restock monitoring will take place. If variants are available, they will be added to the cart.

* Consolidate Parser Error Handling

This commit refactors the monitor to add a new method for handling parser errors. This allows the method to be applied more universally to the monitor as well as the restock monitor without code duplication.

* Update status messages to match

This commit updates the status messages for restocks so they match the ones emitted from the checkout class.

* Add Restocking Handler to Task Runner

This commit adds the restock monitor and restocking handler to the TaskRunner. All nextStates that deal with restocking have been updated to use the restocking monitor.

* Fix Imports

Fix imports to be direct instead of inside an export object.

Co-Authored-By: pr1sm <dhanwada.dev@gmail.com>

* Fix Typo and Reword Status Message

This commit updates a typo that caused a crash when monitoring with keywords. This commit also updates the message that gets emitted when a product has been restocked.

* Bug Fixes

This commit updates several files with the following fixes:
- Fixed typos where the wrong variable was referenced
- Returned the right value
- Fixed Typos in status messages and log messages
- Ensure product url is always the correct format

* Wait for Monitor Delay when Restocking

Co-Authored-By: pr1sm <dhanwada.dev@gmail.com>

Fixes #380