Add Small Frontend Improvements (#395)
* Update yarn lock

* Remove UK Mens Size Group

This commit removes the UK mens size group until multi region size support is implement. The US sizing is renamed to just "Men's" instead of "US Men's"

* Change default country to US

This commit adds a new migrator to adjust the default state of all locations to the US. If a country was already entered, it is not changed.

* remove isMulti from sizes

* Add Better Handling for Garment Size Matching

This commit updates the variant matching algorithm to improve matching when garment sizes are the input.

When a garment size is determined (no numbers in the input size), All mapped sizes must be non-numeric and start with the input size.

* Increase Test Coverage

This commit updates tests related to changed code to increase coverage.

* Fix Size Change Handlers

This commit updates the size change handlers for both the createTask and taskRow components to allow for single select components. A temporary hot fix to mock two multi-select changes is used to prevent a total rewrite of reducer logic.

* Update packages/frontend/src/__tests__/constants/getAllSizes.test.js

* Fix Typo