Generalize Runner Type Tag
This commit updates the TaskRunner to add a type property. A set of constants has been defined for further use if needed.

Now, payloads emitted from the runner will include it's type. This prevents the need for the ShippingRatesRunner to inject it in using a method override.

Finally, a special "done" flag is emitted when the task runner stops. This will allow the frontend to easily determine when the runner has stopped.