Add Support for Starting the ShippingRatesRunner
This commit updates the managers, as well as the runner scripts to create an instances of the ShippingRatesRunner based on an incoming type. This allows TaskManagers to start ShippingRatesRunners.

The change is backwards compatible and passing no specific type will by default spawn a TaskRunner. This means that the current usage in the Frontend is unaffected.